+++ python/src/trezorlib/tools.py
@@ -21,7 +21,6 @@ import struct
 import unicodedata
 from typing import List, NewType
 
-from .coins import slip44
 from .exceptions import TrezorFailure
 
 CallException = TrezorFailure
@@ -168,11 +167,6 @@ def parse_path(nstr: str) -> Address:
     if n[0] == "m":
         n = n[1:]
 
-    # coin_name/a/b/c => 44'/SLIP44_constant'/a/b/c
-    if n[0] in slip44:
-        coin_id = slip44[n[0]]
-        n[0:1] = ["44h", "{}h".format(coin_id)]
-
     def str_to_harden(x: str) -> int:
         if x.startswith("-"):
             return H_(abs(int(x)))
@@ -183,8 +177,8 @@ def parse_path(nstr: str) -> Address:
 
     try:
         return [str_to_harden(x) for x in n]
-    except Exception:
-        raise ValueError("Invalid BIP32 path", nstr)
+    except Exception as e:
+        raise ValueError("Invalid BIP32 path", nstr) from e
 
 
 def normalize_nfc(txt):