+++ python/CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 _At the moment, the project does **not** adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). That is expected to change with version 1.0._
 
-## [0.12.0] - Unreleased
+## [0.12.0] - 2020-04-01
+[0.12.0]: https://github.com/trezor/trezor-firmware/compare/python/v0.11.6...python/v0.12.0
 
 ### Added
 
@@ -398,8 +399,10 @@ _At the moment, the project does **not** adhere to [Semantic Versioning](https:/
 [#352]: https://github.com/trezor/python-trezor/issues/352
 [f#41]: https://github.com/trezor/trezor-firmware/issues/41
 [f#87]: https://github.com/trezor/trezor-firmware/issues/87
+[f#116]: https://github.com/trezor/trezor-firmware/issues/116
 [f#117]: https://github.com/trezor/trezor-firmware/issues/117
 [f#224]: https://github.com/trezor/trezor-firmware/issues/224
+[f#226]: https://github.com/trezor/trezor-firmware/issues/226
 [f#363]: https://github.com/trezor/trezor-firmware/issues/363
 [f#411]: https://github.com/trezor/trezor-firmware/issues/411
 [f#420]: https://github.com/trezor/trezor-firmware/issues/420
@@ -410,3 +413,4 @@ _At the moment, the project does **not** adhere to [Semantic Versioning](https:/
 [f#680]: https://github.com/trezor/trezor-firmware/issues/680
 [f#681]: https://github.com/trezor/trezor-firmware/issues/681
 [f#778]: https://github.com/trezor/trezor-firmware/issues/778
+[f#823]: https://github.com/trezor/trezor-firmware/issues/823